[WIP] factor out trace functionality into separate package and add Go docs
