+++ daemon/logging/trace/trace_unique_concurrent_task_namer_test.go
@@ -1,4 +1,4 @@
-package logging
+package trace
 
 import (
 	"fmt"
